requirements.txt
numpy
scipy
matplotlib
pyqt
nidaqmx
